pyproject.toml
[project]
name = "agent_algorithms"
version = "0.1.0"
description = "Proyecto de agentes y algoritmos."
authors = ["Jhonder18 <jhonder@example.com>"]
requires-python = ">=3.8"

dependencies = [
  "annotated-doc==0.0.4",
  "annotated-types==0.7.0",
  "anyio==4.11.0",
  "blockbuster==1.5.25",
  "cachetools==6.2.1",
  "certifi==2025.10.5",
  "cffi==2.0.0",
  "charset-normalizer==3.4.4",
  "click==8.3.0",
  "cloudpickle==3.1.2",
  "colorama==0.4.6",
  "cryptography==44.0.3",
  "fastapi==0.121.1",
  "filetype==1.2.0",
  "forbiddenfruit==0.1.4",
  "google-ai-generativelanguage==0.9.0",
  "google-api-core==2.28.1",
  "google-api-python-client==2.187.0",
  "google-auth==2.43.0",
  "google-auth-httplib2==0.2.1",
  "googleapis-common-protos==1.72.0",
  "grpcio==1.76.0",
  "grpcio-status==1.76.0",
  "grpcio-tools==1.75.1",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httplib2==0.31.0",
  "httptools==0.7.1",
  "httpx==0.28.1",
  "idna==3.11",
  "importlib_metadata==8.7.0",
  "jsonpatch==1.33",
  "jsonpointer==3.0.0",
  "jsonschema_rs==0.29.1",
  "langchain==1.0.5",
  "langchain-core==1.0.4",
  "langchain-google-genai==3.0.2",
  "langgraph==1.0.3",
  "langgraph-api==0.5.12",
  "langgraph-checkpoint==3.0.1",
  "langgraph-cli==0.4.7",
  "langgraph-prebuilt==1.0.2",
  "langgraph-runtime-inmem==0.18.0",
  "langgraph-sdk==0.2.9",
  "langsmith==0.4.42",
  "lark==1.1.9",
  "opentelemetry-api==1.38.0",
  "opentelemetry-exporter-otlp-proto-common==1.38.0",
  "opentelemetry-exporter-otlp-proto-http==1.38.0",
  "opentelemetry-proto==1.38.0",
  "opentelemetry-sdk==1.38.0",
  "opentelemetry-semantic-conventions==0.59b0",
  "orjson==3.11.4",
  "ormsgpack==1.12.0",
  "packaging==25.0",
  "pip==25.3",
  "proto-plus==1.26.1",
  "protobuf==6.33.0",
  "pyasn1==0.6.1",
  "pyasn1_modules==0.4.2",
  "pycparser==2.23",
  "pydantic==2.12.4",
  "pydantic_core==2.41.5",
  "PyJWT==2.10.1",
  "pyparsing==3.2.5",
  "python-dotenv==1.2.1",
  "PyYAML==6.0.3",
  "requests==2.32.5",
  "requests-toolbelt==1.0.0",
  "rsa==4.9.1",
  "setuptools==65.5.0",
  "sniffio==1.3.1",
  "sse-starlette==2.1.3",
  "starlette==0.49.3",
  "structlog==25.5.0",
  "tenacity==9.1.2",
  "tqdm==4.67.1",
  "truststore==0.10.4",
  "typing_extensions==4.15.0",
  "typing-inspection==0.4.2",
  "uritemplate==4.2.0",
  "urllib3==2.5.0",
  "uvicorn==0.38.0",
  "watchfiles==1.1.1",
  "websockets==15.0.1",
  "xxhash==3.6.0",
  "zipp==3.23.0",
  "zstandard==0.25.0"
]